### 2025-09-30 #9101 @kyleniemeyer
@loeffko there are some issues with your submission, if you could address these:
- I cannot see where your paper is located or in which branch. Can you please point to this?
- I do not see any installation instructions
- I do not see any tests
- overall the package is structured strangely, and does not appear like a typical software package - I also see a lot of data files (*.gms) that probably do not belong at the top level of the repository
