### 2025-10-06 #9101 @kyleniemeyer
@loeffko ok, we will need to point to the repository holding the paper so that it can be built.

Is it fair to say that the entire https://github.com/GENeSYS-MOD org is under review?

I can see a test directory in the Julia version repo; are there other tests? Either way, the proper way to run the tests needs to be described somewhere clearly, unless CI is set up to do this automatically. I can't find those instructions in the documentation; are they present or does that need to be added?
