### 2025-10-13 03e6cdd
removed erroneous tab characters from paper.md file
