### 2025-12-05 #9101 @mwetz
Hey @marlinarnz - I would be happy to review as I am also familiar with GAMS, Julia and energy system models in general.

For full disclosure: @loeffko reviewed the JOSS publication of our energy system model REMix (https://github.com/openjournals/joss-reviews/issues/6330), which he also mentioned above. If this is considered a potential conflict of interest under JOSS guidelines let me know.
