### 2025-12-06 #9101 @marlinarnz
Thank you very much @Sush1090 and @mwetz for agreeing to review so swiftly!

@Sush1090 : In this case, it is not a problem that you don't have experience with GAMS since it is proprietary software. As of our [reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#what-about-submissions-that-rely-upon-proprietary-languages-development-environments) you may focus on the other parts of the review, as long as you can install the software and verify its functionality. You may suggest open-source alternatives to the authors, if applicable.

I will open a dedicated thread for the reviews beginning of next week.
