### 2025-12-08 #9571 @editorialbot
[REVIEW]: The Global Energy System Model (GENeSYS-MOD) v4.0 - A Flexible Energy System Modelling Framework for Julia and GAMS
**Submitting author:** @loeffko (<a href="http://orcid.org/0000-0002-5435-1880">Konstantin Löffler</a>)
**Repository:** https://github.com/GENeSYS-MOD/GENeSYS_MOD.readthedocs
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.5
