### 2026-01-14 #9571 @mwetz
- [x]  The paper URL links to the documentation page, where further linking to the two implementations in GAMS and Julia are provided. I suggest the authors provide additional hyperlinks in the README of the readthedocs repo to link to the two implementations and their respective documentations directly.
  - [x] The chosen license (Appache License 2.0) is OSI approved
  - [x] The Julia repository includes a LICENSE file, but is missing the license holder and year
  - [x] The GAMS repository includes both a LICENSE file and a LICENSE.txt file. The file without file extension is missing the license holder information, the .txt file correctly states the copyright holder.
  - [x] Yes, both the submitting author and the authors team are in line with the repositories provided
  - [x] Yes, the tool is used across multiple research institutions and was used for a range of different academic publications in journals and conferences

  - [x] Julia installation instructions should be improved to properly install the package and run a sample case
  - [x] Examples for both the GAMS and Julia implementation are both running

  - [x] Description of core functionality, variables and parameters is sufficient
  - [x]  A short description high level options of the genesys call for the Julia implementation should be included in the documentation
  - [x] (Ideally, the mathematical formulation from the Model should be listed in the documentation as well)
  - [x] Julia implementation includes proper testing, the GAMS implementation does not include proper testing, but multiple examples using sample data are included.

- [x] Minor issues in the software paper
  - [x] Paper is written in American English, but uses British English for "modelling"
  - [x] Line 34: including GENeSYS-MOD **,** can be found (missing comma)
  - [x] Line 123: Overview ~in~ **on** the respective strengths
  - [x] Line 132:  in that regard, ~like~ **as** described in
