### 2026-01-14 7c68b52
Update README.rst
