+++ README.rst
@@ -1,4 +1,23 @@
-Source files for the GENeSYS-MOD readthedocs page
+GENeSYS-MOD.readthedocs
 =======================================
 
-Read the tutorial here: https://genesysmod.readthedocs.io/en/latest/
+This repository contains the source code for the readthedocs tutorial page (see: https://genesysmod.readthedocs.io/en/latest/).
+
+The readthedocs includes information on installation procedures, some examples, as well as background information on mathematical formulation and structure. 
+
+
+Links to model repositories
+-------------------------------------------------
+
+For the individual model repositories please refer to the following pages:
+
+GENeSYS-MOD.gms (GAMS-based version): https://github.com/GENeSYS-MOD/GENeSYS_MOD.gms 
+
+GENeSYS-MOD.jl (Julia-based version): https://github.com/GENeSYS-MOD/GENeSYSMOD.jl
+
+Further repositories
+--------------------
+
+For raw input data used in research projects and studies, please refer to https://github.com/GENeSYS-MOD/GENeSYS_MOD.data
+
+For verious tools around GENeSYS-MOD, please refer to https://github.com/GENeSYS-MOD/GENeSYS_MOD.tools
