### 2026-01-14 ccc69bb
fixed issue with DOI of one reference
