+++ paper/paper.bib
@@ -177,7 +177,7 @@
 	volume = {20},
 	issn = {2211-467X},
 	url = {https://www.sciencedirect.com/science/article/pii/S2211467X18300142},
-	doi = {https://doi.org/10.1016/j.esr.2018.03.005},
+	doi = {10.1016/j.esr.2018.03.005},
 	journal = {Energy Strategy Reviews},
 	author = {Gardumi, Francesco and Shivakumar, Abhishek and Morrison, Robbie and Taliotis, Constantinos and Broad, Oliver and Beltramo, Agnese and Sridharan, Vignesh and Howells, Mark and Hörsch, Jonas and Niet, Taco and Almulla, Youssef and Ramos, Eunice and Burandt, Thorsten and Balderrama, Gabriela Peña and Moura, Gustavo Nikolaus Pinto de and Zepeda, Eduardo and Alfstad, Thomas},
 	year = {2018},
