### 2026-01-15 a7400b7
removed placeholder for GENeSYS-MOD.jl release notes added some more clarification regarding versioning
