### 2026-02-10 d1bba72
updated Julia quickstart guide after updates in the model repo
