+++ README.rst
@@ -1,4 +1,15 @@
 Source files for the GENeSYS-MOD readthedocs page
 =======================================
 
-Read the tutorial here: https://genesysmod.readthedocs.io/en/latest/
+Documentation
+~~~~~~~~~~~~~
+
+Read the documentation here: https://genesysmod.readthedocs.io/en/latest/
+
+
+Model repositories
+~~~~~~~~~~~~~~~~~~
+
+Find the model repositories here:
+**GAMS**: https://github.com/GENeSYS-MOD/GENeSYS_MOD.gms
+**Julia**: https://github.com/GENeSYS-MOD/GENeSYSMOD.jl
