### 2026-02-15 25502d9
added table of fuels in GENeSYS-MOD closes #18
