+++ docs/source/mathematical_formulation/units_and_data.rst
@@ -33,3 +33,73 @@ List of fuels
 
 The following fuels are included in the current main version of GENeSYS-MOD (other fuels might be included in specific versions used for case studies etc.).
 
+.. list-table::
+   :header-rows: 1
+   :widths: 28 72
+
+   * - Name
+     - Description
+   * - Air
+     - Dummy fuel used to calculate CO2 contents of air in DAC
+   * - Area_Rooftop_Commercial
+     - Surface area for rooftop installations (commercial buildings)
+   * - Area_Rooftop_Residential
+     - Surface area for rooftop installations (residential buildings)
+   * - Biofuel
+     - Biofuels (e.g. bio-diesel)
+   * - Biomass
+     - Solid biomass
+   * - ETS
+     - Dummy fuel used to emulate an emission-trading-system
+   * - ETS_Source
+     - Dummy fuel used to emulate an emission-trading-system
+   * - Gas_Bio
+     - Methane (from biological origin)
+   * - Gas_Natural
+     - Methane (from geological origin)
+   * - Gas_Synth
+     - Methane (from synthetic origin / hydrogen-derivative)
+   * - H2
+     - Hydrogen (gaseous)
+   * - H2_Blend
+     - Hydrogen-methane blend
+   * - Hardcoal
+     - Steam / hard coal
+   * - Heat_Buildings
+     - Space heating and warm water in buildings
+   * - Heat_District
+     - Heat feed-in into the district heating network
+   * - Heat_High_Industrial
+     - High-temperature industrial process heat (>1000°C)
+   * - Heat_Low_Industrial
+     - Low-temperature industrial process heat (<100°C)
+   * - Heat_MediumLow_Industrial
+     - Medium-temperature industrial process heat (100-400°C)
+   * - Heat_MediumHigh_Industrial
+     - Medium-temperature industrial process heat (400-1000°C)
+   * - LBG
+     - Liquid biogas / liquid form of Gas_Bio
+   * - LH2
+     - Liquid hydrogen (H2)
+   * - Lignite
+     - Lignite coal
+   * - LNG
+     - Liquefied natural gas / liquid form of Gas_Natural
+   * - LSG
+     - Liquefied synthetic gas / liquid form of Gas_Synth
+   * - Mobility_Freight
+     - Freight mobility services
+   * - Mobility_Passenger
+     - Passenger mobility services
+   * - Nuclear
+     - Nuclear fuel (e.g. fuel rods)
+   * - Oil
+     - Petroleum products
+   * - Power
+     - Electricity / power
+   * - Powerfuel
+     - E-fuels (e.g. synthetically derived diesel or gasoline)
+   * - Waste
+     - Waste energy resource
+
+
