### 2026-02-15 c287a7d
added list of technologies currently being used in GENeSYS-MOD v4 closes #17
