+++ docs/source/mathematical_formulation/units_and_data.rst
@@ -27,6 +27,334 @@ List of technologies
 
 The following technologies are included in the current main version of GENeSYS-MOD (other technologies might be included in specific versions used for case studies etc.). Please refer to :ref:`the set documentation <technology>` for an explanation of the naming convention.
 
+.. list-table::
+   :header-rows: 1
+   :widths: 28 72
+
+   * - Name
+     - Description
+   * - A_Air
+     - Dummy fuel for use in Direct Air Capture (DAC)
+   * - A_Rooftop_Commercial
+     - Surface area of commercial rooftops
+   * - A_Rooftop_Residential
+     - Surface area of residential rooftops
+   * - CHP_Biomass_Solid
+     - Combined-heat-and-power plant using solid biomass
+   * - CHP_Biomass_Solid_CCS
+     - Combined-heat-and-power plant using solid biomass with CCS
+   * - CHP_Coal_Hardcoal
+     - Combined-heat-and-power plant using hard coal
+   * - CHP_Coal_Hardcoal_CCS
+     - Combined-heat-and-power plant using hard coal with CCS
+   * - CHP_Coal_Lignite
+     - Combined-heat-and-power plant using lignite coal
+   * - CHP_Coal_Lignite_CCS
+     - Combined-heat-and-power plant using lignite coal with CCS
+   * - CHP_Gas_CCGT_Biogas
+     - Combined-heat-and-power plant using biogas
+   * - CHP_Gas_CCGT_Biogas_CCS
+     - Combined-heat-and-power plant using biogas with CCS
+   * - CHP_Gas_CCGT_Natural
+     - Combined-heat-and-power plant using natural gas
+   * - CHP_Gas_CCGT_Natural_CCS
+     - Combined-heat-and-power plant using natural gas with CCS
+   * - CHP_Gas_CCGT_SynGas
+     - Combined-heat-and-power plant using synthetic natural gas
+   * - CHP_Hydrogen_FuelCell
+     - Combined-heat-and-power plant using hydrogen in a fuel cell
+   * - CHP_Oil
+     - Combined-heat-and-power plant using crude oil
+   * - CHP_WasteToEnergy
+     - Combined-heat-and-power plant using waste
+   * - D_Battery_Li-Ion
+     - Storage dummy technology / grid link for Lithium-Ion Batteries
+   * - D_Battery_Redox
+     - Storage dummy technology / grid link for Redox-Flow Batteries
+   * - D_CAES
+     - Storage dummy technology / grid link for Compressed-Air Electricity Storage
+   * - D_Gas_H2
+     - Storage dummy technology / grid link for Redox-Flow Batteries
+   * - D_Gas_Methane
+     - Storage dummy technology / grid link for Redox-Flow Batteries
+   * - D_HB_Tank_Small
+     - Storage dummy technology / grid link for Redox-Flow Batteries
+   * - D_HD_Pit
+     - Storage dummy technology / grid link for large-scale heat storage (e.g. aquifer)
+   * - D_HLI_Tank_Large
+     - Storage dummy technology / grid link for Redox-Flow Batteries
+   * - D_PHS
+     - Storage dummy technology / grid link for Redox-Flow Batteries
+   * - FRT_Rail_Conv
+     - Freight railroad technology using conventional (petro-) fuels
+   * - FRT_Rail_Electric
+     - Freight railroad technology using electricity
+   * - FRT_Road_BEV
+     - Freight trucks using battery-electric technology
+   * - FRT_Road_H2
+     - Freight trucks using hydrogen in a fuel cell
+   * - FRT_Road_ICE
+     - Freight trucks with an internal combustion engine, can use multiple fuels
+   * - FRT_Road_LNG
[…]
