### 2026-02-15 5e963d2
added some more papers, removed one wrong reference prepared version for 0.3.5
