+++ docs/source/references.bib
@@ -317,21 +317,6 @@
 	pages = {124122},
 }
 
-@article{shen_environmental_2024,
-	title = {Environmental and climate impacts of a large-scale deployment of green hydrogen in {Europe}},
-	volume = {5},
-	issn = {26662787},
-	url = {https://linkinghub.elsevier.com/retrieve/pii/S2666278724000096},
-	doi = {10.1016/j.egycc.2024.100133},
-	language = {en},
-	urldate = {2025-02-08},
-	journal = {Energy and Climate Change},
-	author = {Shen, Haiping and Crespo Del Granado, Pedro and Jorge, Raquel Santos and Löffler, Konstantin},
-	month = dec,
-	year = {2024},
-	pages = {100133},
-}
-
 @article{barani_road_2025,
 	title = {The road to a sustainable energy system in the {Guadeloupe} archipelago: {Challenges} and opportunities},
 	volume = {11},
@@ -539,4 +524,36 @@
 	month = may,
 	year = {2022},
 	pages = {112174},
+}
+
+@article{barani_european_2026,
+	title = {European energy vision 2050 and beyond: {Designing} scenarios for {Europe}’s energy transition},
+	volume = {225},
+	issn = {13640321},
+	shorttitle = {European energy vision 2050 and beyond},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S1364032125007476},
+	doi = {10.1016/j.rser.2025.116074},
+	language = {en},
+	urldate = {2026-02-15},
+	journal = {Renewable and Sustainable Energy Reviews},
+	author = {Barani, Mostafa and Löffler, Konstantin and Crespo Del Granado, Pedro and Moskalenko, Nikita and Panos, Evangelos and Hoffart, Franziska M. and Hirschhausen, Christian Von and Kannavou, Maria and Auer, Hans and Hainsch, Karlo and González Grandón, Tatiana and Mathisen, Siri and Tomasgard, Asgeir},
+	month = jan,
+	year = {2026},
+	pages = {116074},
+}
+
+@inproceedings{mathisen_energy_2025,
+	address = {Lisbon, Portugal},
+	title = {Energy and {Climate} {Plans} in {Energy} {System} {Modelling} {Scenarios}},
+	copyright = {https://doi.org/10.15223/policy-029},
+	isbn = {979-8-3315-1278-1},
+	url = {https://ieeexplore.ieee.org/document/11050155/},
+	doi = {10.1109/EEM64765.2025.11050155},
+	urldate = {2026-02-15},
+	booktitle = {2025 21st {International} {Conference} on the {European} {Energy} {Market} ({EEM})},
+	publisher = {IEEE},
+	author = {Mathisen, Siri and Löffler, Konstantin and Barani, Mostafa and Belsnes, Michael M.},
+	month = may,
+	year = {2025},
+	pages = {1--6},
 }
