### 2026-02-15 ad9c39d
added list of switches to be chosen in GENeSYS-MOD julia implementation could still use a short glance if this is also correct, plus some implementation info
