+++ docs/source/how_to_guides/gams/user_guide/switches.rst
@@ -3,4 +3,232 @@ Switches
 
 .. _switches:
 
-*Coming soon. :)*
+The list below provides a list of all available settings / switches for GENeSYS-MOD that are given to the user as a default.
+
+In GAMS, the switches can either be set via the genesysmod.gms file (where they are located), but can also be added via the command line, either in GAMS Studio / IDE (add ``--switch_name=option`` to the command line), or in a terminal / batch-file, by adding ``--switch_name=option`` to your call of gams.
+Setting any option from outside of genesysmod.gms will overwrite the default settings located there.
+
+.. list-table::
+   :header-rows: 1
+   :widths: 26 14 8 48
+
+   * - Name
+     - Options
+     - Default
+     - Description
+   * - **General settings**
+     -
+     -
+     -
+   * - data_file
+     - *[string]*
+     -
+     - The data file to use for the calculations (can be overwritten if using a pre-defined scenario option)
+   * - hourly_data_file
+     - *[string]*
+     -
+     - The timeseries file to use for the calculations (can be overwritten if using a pre-defined scenario option)
+   * - model_region
+     - *[string]*
+     -
+     - Name of the modeled region (relevant for the use of a region-specific scenario data file)
+   * - data_base_region
+     - *[string]*
+     -
+     - Name of the base region of the modeled region (the base region will be used to populate missing data)
+   * - year
+     - *[string]*
+     - 2018
+     - The start-year of the modelling horizon
+   * - emissionPathway
+     - *[string]*
+     -
+     - Name of the scenario / pathway to compute
+   * - emissionScenario
+     - *[string]*
+     -
+     - Additional scenario information, e.g. for distribution purposes
+   * - **Timeseries reduction**
+     -
+     -
+     -
+   * - timeseries_method
+     - *[elmod]*
+     - elmod
+     - Chooses the timeseries reduction method. Currently, only elmod is implemented
+   * - elmod_nthhour
+     - *[integer]*
+     - 484
+     - Defined the step-size for the timeseries reduction algorithm. Good choices include 784, 484, 244, 122, 73, 25. Lower numbers will lead to more time slices and more computational requirements
+   * - elmod_starthour
+     - *[integer]*
+     - 8
+     - Determines the start-hour of the timeseries reduction algorithm
+   * - elmod_dunkelflaute
+     - *[0,1]*
+     - 0
+     - Enables / disables injection of an artificial "Dunkelflaute" into the timeseries
+   * - **Data input/output**
+     -
+     -
+     -
+   * - switch_test_data_load
+     - *[0,1]*
+     - 0
+     - If enabled, the code will just perform a test loading of all data and terminate before the equations are constructed
+   * - switch_only_load_gdx
+     - *[0,1]*
+     - 0
+     - If enabled, the model will just read in the data in .gdx format and not attempt to read from the Excel file
+   * - switch_unixPath
+     - *[0,1]*
+     - 0
+     - Determines the structure of the directory paths. Needs to be 1 for Unix-based systems (e.g. Mac OS, Linux) and 0 for Windows
+   * - switch_read_data_long
+     - *[0,1]*
+     - 1
+     - Determines the data structure, the codes assumes to read. 1 assumes tables in long format, 0 assumes tables in wide format
+   * - switch_write_output
+     - *[gdx,csv,xls]*
+     - gdx
+     - Type of output files to be created. For single runs, csv or xls is recommended. If multiple runs should be compared, using .gdx and then merging it with the script in the results folder is recommended
+   * - switch_only_write_results
+     - *[0,1]*
+     - 0
+     - Reads in a result gdx file and outputs the results as e.g. xls without re-computation
[…]
