### 2026-02-25 #9571 @Sush1090
Hello @marlinarnz

I am only waiting for the Julia documentation to be updated. It needs to be updated with the usage of the registered package.
Issue tracker : [link](https://github.com/GENeSYS-MOD/GENeSYSMOD.jl/issues/34)

Apart from this most of it is done from my side.
