### 2026-02-25 e4a668d
changed minor issues recommended by @mwetz from the JOSS review
