### 2026-02-25 #9571 @loeffko
Hi, @mwetz , sorry that I did not see your minor remarks on the paper manuscript before. I have just pushed a new version to https://github.com/GENeSYS-MOD/GENeSYS_MOD.readthedocs/tree/joss

You can find the updated pdf here: https://github.com/GENeSYS-MOD/GENeSYS_MOD.readthedocs/actions/runs/22406609632/artifacts/5657670762

Edit:

Also, a list of available settings / switches has been added for both GAMS and Julia in the readthedocs, as well as a full list of regularly included technologies and fuels. https://genesysmod.readthedocs.io/en/latest/how_to_guides/julia/user_guide/switches.html
