### 2026-02-26 #9571 @mwetz
Hi @loeffko, thanks for updating the paper and also the documentation on the switches. This also addresses the final open issue regarding additional information on how to call the Julia function.

Hi @marlinarnz,  from my side, all the issues have been resolved and the publication can proceed. The only slightly open point is the mathematical description of the model, but, speaking from personal experience, I know that can require a significant amount of effort. However, since I don't want this to delay the otherwise excellent publication, I would treat it more as encouragement for the authors to add to the documentation in future.
