### 2026-02-27 8114d78
improved quick-start guide for GENeSYS-MOD.jl following the JOSS review comments separated the more advanced / development installation from the quick start guide
