+++ docs/source/how_to_guides/julia/quick_start/quick-start.md
@@ -17,25 +17,9 @@
   *	[CPLEX](https://www.ibm.com/products/ilog-cplex-optimization-studio/cplex-optimizer)
   *	[HiGHS](https://highs.dev/), open source solver.
 
-5.	[Optional] This step is optional and can be skipped if you only want to run the model and do not need to look into the code behind the model. If you choose to clone, follow steps 6 and 7 as well, otherwise go to step 8. Clone the GitHub repository for [GENeSYSMOD.jl](https://github.com/GENeSYS-MOD/GENeSYSMOD.jl)
-  *	Download git: [Git - Downloads](https://git-scm.com/)
-  *	Navigate to the folder where you want the repo to be located.
-  * Open Git Bash by right clicking in the chosen folder and choosing "Git Bash Here"
-  * Type the following command in Git Bash:
+5.  Open the Julia REPL by using the Alt+J+O command. The REPL(read-eval-print loop) is the interactive command line interface in Julia.
 
-```
-git clone https://github.com/GENeSYS-MOD/GENeSYSMOD.jl.git
-git pull
-```
-
-6.	[Optional] Open GENeSYS-MOD folder in Visual Studios
-  *	File > Open folder
-
-7.	[Optional] Change to Julia environment for GENeSYSMOD.jl.
-
-8.  Open the Julia REPL by using the Alt+J+O command. The REPL(read-eval-print loop) is the interactive command line interface in Julia.
-
-9.	Install packages in Visual Studio:
+6.	Install packages in Visual Studio:
   *	In the REPL terminal, type ']' to change to enter pkg mode
   * Type 'add <package>', where package is equal to:
     - GENeSYSMOD
@@ -45,29 +29,66 @@ git pull
     - DataFrames
     - CSV	
     - Ipopt
-    - <your chosen solver>
+    - *(your chosen solver)*
 
-10.	You have now installed all packages necessary to run GENeSYS-MOD and can run a case by using the genesysmod function. Examples can be found in the tests.
+7.	You have now installed all packages necessary to run GENeSYS-MOD and can run a case by using the genesysmod function. Examples can be found in the tests.
 
 ## Running a case
 
 Create a folder where you will be able to store your scripts, inputs and outputs of the model. 
 
-If you have cloned the repository and plan on changing the code of the model locally, you can then include the following at the top of your script: 
-
-```julia
-Pkg.develop(path="..\\GENeSYSMOD.jl")
-using GENeSYSMOD
-```
-Where you replace the path with the relative or absolute path to the package. Note the "\\".
-
-Otherwise (**RECOMMENDED**), for simply running the model without modifying the model or constraints, use:
+For simply running the model without modifying the model or constraints, use:
 ```julia
 using GENeSYSMOD
 ```
 
-You can then run the model by running the function `genesysmod` with the appropriate `Switch`.
+You can then run the model by running the function `genesysmod` with the appropriate `Switches` (see documentation here: [Switches](/how_to_guides/julia/user_guide/switches)).
 
+## Example script
+
+You can use the following example script to then run GENeSYS-MOD with the open solvers HiGHS and Ipopt:
+
+```julia
+using GENeSYSMOD
+using HiGHS
+using Ipopt
+
+solver = HiGHS.Optimizer
+
+const TEST_RESULTS_DIR = joinpath(pkgdir(GENeSYSMOD),"test","TestData","Results")
+
+model, data = genesysmod(;elmod_daystep = 80, elmod_hourstep = 4, solver=solver, DNLPsolver = Ipopt.Optimizer, threads=0,
[…]
