### 2026-02-27 #9571 @loeffko
Hi @mwetz what would you envision there? An actual equation by equation documentation? Or would e.g. a description of each equation block (e.g. all EB constraints which are the energy balances) suffice? Some key equations could of course be singled out, but I feel like a full list of all individual equations is a bit cumbersome (especially since we already use speaking parameter and variable names, so it should be mostly clear and "human-readable").

@Sush1090 thanks for reminding us of that issue! I have just pushed an update to our readthedocs development branch: https://genesysmod.readthedocs.io/en/development/how_to_guides/julia/quick_start/quick-start.html
I will also post this to your issue, please let me know if that makes it better/clearer. :)
