### 2026-02-27 #9571 @Sush1090
Hello @loeffko and @marlinarnz
Thank you for addressing the issue. From my end all the aspects are clear and met according to guidelines.
As the code is already part Open Energy Modelling Initiative : [link](https://wiki.openmod-initiative.org/wiki/GENeSYS-MOD), I have no additional comments.

Excellent package!!
