### 2026-03-06 #9571 @mwetz
@loeffko Fair point, it certainly depends how strongly abstracted the equations are since a long list with repeated equations also does not help. Listing some selected equations to convey the concepts could be a compromise to help new users to grasp the mathematical foundation behind the model. Also with Julia it should be quite straightforward to host some equations directly based on the code (see e.g. the GenX docs which uses DosStrings for this purpose: https://genxproject.github.io/GenX.jl/stable/Model_Reference/Resources/curtailable_variable_renewable/). From personal experience I cannot recommend doing this for GAMS, since this involves a lot of overhead to convert the source files.
