### 2026-03-19 #9571 @marlinarnz
@loeffko congratulations to your paper! It is great to see GENeSYS-MOD evolving - you and all contributers have put a lot of work in it.

Now, there are two last steps for you to address before acceptance:

- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. Since the paper refers to the readthedocs repository, I would use its version number. The title of your paper (v4.0) won't be affected.
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. Make sure that the license listed for the archive is the same as the software license. Since the paper concerns the entire GENeSYS-MOD ecosystem, it might make sense to deposit the content of all repositories in one static archive. Feel free to suggest a different approach, if you think that would be inappropriate.
