### 2026-03-23 #9571 @loeffko
> [@loeffko](https://github.com/loeffko) I was waiting for a reaction of yours to the last comment addressed to you. I guess thumbs up means that you're not considering any further changes? In this case the reviews are completed and we can progress in the process.

@marlinarnz Hi, yes, sorry for the confusion! We have taken the suggestions and will definitely consider this for a future update of the documentation, but would like to proceed with the status as-is for now. :) Your suggestions sound great, I will create a joint statis Zenodo repo for all sub-repositories.
