### 2026-03-23 #9571 @loeffko
@marlinarnz Hi, so the release number for the readthedocs repo is [0.3.5](https://github.com/GENeSYS-MOD/GENeSYS_MOD.readthedocs/releases/tag/v0.3.5).
Also, the DOI from Zenodo is: [10.5281/zenodo.19187071](https://zenodo.org/records/19187071).

Let me know if this is fine :)
