### 2026-03-31 #9571 @kyleniemeyer
Hi @loeffko, just a few things need to be addressed before accepting:

- in the "New features of GENeSYS-MOD version 4" section, it looks like the numbered list is not formatted correctly as intended. It likely needs an empty line before/after the list.
- In the Zenodo archive, please edit the author list (rather, replace the single author) to match the paper author list.
